this is now a wrapper around the doi plugin
Scraping Amazon has been proved futile. This is now a simple wrapper
around the DOI plugin. It will use the DOI plugin's configured ISBN
resolver to fetch data about the given ASIN, but then still link to
Amazon.

Note that this only works for books because for books, the ASIN is the
ISB-10 of the book.